+++ src/com/dmdirc/QueryFactory.java
@@ -23,7 +23,7 @@
 package com.dmdirc;
 
 import com.dmdirc.interfaces.CommandController;
-import com.dmdirc.ui.core.BackBufferFactory;
+import com.dmdirc.ui.messages.BackBufferFactory;
 import com.dmdirc.ui.input.TabCompleterFactory;
 import com.dmdirc.ui.messages.sink.MessageSinkManager;
 import com.dmdirc.util.URLBuilder;
